Clean yum metadata after versionlock file update
Currently when a versionlock entry is updated it has the unfortunate
problem that the cached yum repository metadata only contains
the previous locked files.

The class `yum::plugin::versionlock` now has a new option `clean`
defaulting to `false`. If true a `yum clean all` will exectuted
after the versionlock file is updated.

A new class `yum::clean` is added as it may be sensible
for some other operations to do the same.