Add support for yum facts (similar to the apt ones) (#141)
* Add facts related to upgradable packages

Add three new facts:
- yum_has_updates -- Boolean indicating if updates are available;
- yum_package_updates -- Array[String] list of upgradable packages;
- yum_updates -- Integer number of upgradable packages.

These facts are named in the same way as the ones provided by the
puppetlabs-apt module.

* Add a new yum_reboot_required fact

Just like the apt_reboot_required fact provided by the puppetlabs-apt
module, this fact tells if the systems needs to be rebooted.  It relies
on the needs-restarting(1) command which is part of yum-utils, already
managed by the module.

* Add 'list updates' as a valid task

This task will run yum without -C, fetching recent data if the
repositories are outdated.  The output is processed to extract the
package name, available version and repository of the update.  This
information is provided as an array of hashes.