Support paramatized yum::versionlock
Motivated by the change of format to the versionlock file with DNF systems
change `yum::versionlock` so it supports parameters `version`,
`release`, `epoch` and `arch`. If version is specified the namevar is now the package name.

So on a DNF based systemd (e.g CentOS 8) the following must be used.

```puppet
yum::versionlock{'bash':
  version => '3.4',
}
```

results in a version lock of on a dnf based system

```
bash-0:3.4-*
```

and on a yum based system of

```
0:bash-3.4-*
```

A fuller example with all parameters set.

```puppet
yum::versionlock{'bash':
  version  => '5.4',
  release  => 'alpha-rc3',
  epoch    => 8,
  arch     => 'noarch'
}
```

results in a version lock on a DNF based system of

```
bash-8:5.4-alpha-rc3.noarch
```

The same mechanism on YUM based systems (e.g CentOS 7) can be used.
To specify this new machanism at least a `version` should be set.

```puppet
yum::versionlock{'bash':
  version  => '5.4',
  release  => 'alpha-rc3',
  epoch    => 8,
  arch     => 'noarch'
  future   => true,
}
```

would results in

```
8:bash-5.4-alpha-rc3.noarch
```

as used by yum. If `version` is left default `undef`  then the old mechanism is used so this
is backwards compatable. e.g

```puppet
yum::versionlock{"0:bash-3.4-7.el8.x86_64"}
```
is still valid

Fixes: 150