Improve extraction of available updates
The `yum list updates` command outputs diagnostic messages and then the
actual available updates both on stdout.  Depending on the installed yum
plugins, a variable number of diagnostic lines are emitted, and the
output is always formatted for humans, making it non-trivial to
distinguish useless and relevant lines.

Instead of ignoring the first line, match each line with a pattern
matching the following fields:
  - package: The name of the package for which an update is available,
    they are in the format <package-name>.<architecture>;
  - available_version: The version of the package available for update,
    ensure it starts with a digit;
  - repository: The repository containing the update

These fields are separated by a variable amount of space.  When a line
match this pattern, consider the package as relevant for update.